==> Step_1_Acquire_Data.ipynb <==
# A&A Project: Sentiment Analysis of Apple M1 in Twitter

Author: Hongshen Lee

Date:  2020/11/21

## **Project Description:**

This project comes from an idea should I buy the new MacBook with new cross-era chips. Apple has released its cross era product with **Apple Silicon M1**. I'm curious about whether this new product is popular and what its reputation is. I want to see people’s attitudes towards the new MacBook products. So, I collect the tweets about this topic and use a sentiment analysis model to predict their attitudes to support my decision. 

The whole project should be divided by three steps:
1. Step 1: Pulling tweets about the products(macbook with new M1 chips) from twitter
2. Step 2: Building a model to do the sentiment analysis (with relevant dataset to train)
3. Step 3: Combining the data and model to predict the sentiment of these tweets, and go to the conclusion.

## Step 1: Pulling Data with Tweepy

This part is to scraped tweets about the topic to form a dataset. I use tweepy to do this, using keywords and api.search() to search for some tweets relevant to my topic. Keywords include apple m1, MBA, MacBook air, apple chip and so on.

But these tweets are not necessarily related to our topic "Apple's new chip", so setting a since time can effectively improve the relevance of tweets to our topic.


import os
import time
import pandas as pd
from datetime import datetime
import tweepy
# I've put my API keys in a .py file called API_keys.py
from API_keys import api_key, api_key_secret, access_token, access_token_secret

### Step 1.1: Authenticate API


# Authenticate the Tweepy API
auth = tweepy.OAuthHandler(api_key,api_key_secret)
auth.set_access_token(access_token, access_token_secret)

# Once the rate limit is hit, we will be notified that we must wait 15 mins (900 secs)
api = tweepy.API(auth,wait_on_rate_limit=True,wait_on_rate_limit_notify=True, compression=True)

### Step 1.2: Search and Store the Data

### Fields
For each record, it conains fields: 

- id_str: unique id for each tweet
- username: twitter's screen name in twitter
- location: twitter's location information in Twitter
- following: the number of people twitter follows
- followers: the number of people following the Twitter
- total_tweers: the number of total tweets published by this Twitter
- favorite_count: the number of the favorite clicks of this tweet
- retweet_count: the number of retweets of this tweet.
- text: the context of this tweet
- source: Twitter uses what to publish this tweet.

### Discussion

For this dataset, serval interesting problems could be:

- How people like or not the new product with new chips from Apple?
- Do active and influential people (with more followers and total tweets) like new products?
- Do people's preferences have anything to do with their geographical location ？
- Do people's preferences have anything to do with their phones (could be inferred by source field) ？


file_name="apple_m1_data_new.csv"

def init_file():
    db_tweets = pd.DataFrame(columns=['id_str','username', 'location', 'following',
                                      'followers', 'total_tweets', 'favorite_count',
                                      'retweet_count', 'text', 'source'])
    db_tweets.to_csv(file_name, index=False)
    

def write_down_records(tweet_list):
    db_tweets = pd.DataFrame(columns=['id_str','username', 'location', 'following',
                                      'followers', 'total_tweets', 'favorite_count',
                                      'retweet_count', 'text', 'source'])
    for tweet in tweet_list:
        id_str = tweet.id_str
        username = tweet.user.screen_name
        location = tweet.user.location
        following = tweet.user.friends_count
        followers = tweet.user.followers_count
        total_tweets = tweet.user.statuses_count
        retweet_count = tweet.retweet_count
        favorite_count = tweet.favorite_count
        try:
            text = tweet.full_text
            source = tweet.source
        except AttributeError:  # A Retweet
            text = tweet.retweeted_status.full_text
            source = tweet.retweeted_status.source
        # Add the 11 variables to the empty list - ith_tweet:
        ith_tweet = [id_str,username, location, following, followers, total_tweets,
                 retweet_count, favorite_count, text, source]
        db_tweets.loc[len(db_tweets)] = ith_tweet
    db_tweets.to_csv(file_name, mode='a', header=False, index=False)

def scrape_tweets(search_words, date_since):
    # Define a for-loop to generate tweets at regular intervals
    # We cannot make large API call in one go. Hence, let's try T times

    program_start = time.time()
    # Collect tweets using the Cursor object
    # .Cursor() returns an object that you can iterate or loop over to access the data collected.
    count=1;
    for tweets in tweepy.Cursor(api.search, q=search_words, lang="en", rpp=100,count=100, 
                        since=date_since, tweet_mode='extended').pages():
        # We will time how long it takes to scrape tweets for each run:
        start_run = time.time()
        # Store these tweets into a python list
        tweet_list = [tweet for tweet in tweets]
        noTweets = len(tweet_list)
        # Wrtie down the records to file
        write_down_records(tweet_list)
        # Run ended:
        duration_run = round((time.time() - start_run) / 60, 2)
        print('Time take for {} to complete is {} mins,scraped {} tweets'.format(count,duration_run, noTweets))
        count=count+1
        # time.sleep(920)  # 15 minute sleep time

    # End
    program_end = time.time()
    print('Scraping has completed!')
    print('Total time taken to scrap is {} minutes.'.format(round(program_end - program_start) / 60, 2))

### Step 1.3: Determine Key Words and Carry out the Search
Some keywords like apple m1, MBA, MacBook air, apple chip to search the tweets. 

But these tweets are not necessarily related to our topic "Apple's new chip", so setting a since time can effectively improve the relevance of tweets to our topic.

init_file()

search_words = "#M1 -filter:retweets"
date_since = "2020-11-10"
# Call the function scraptweets
scrape_tweets(search_words, date_since)

search_words = "MBA OR MACBOOK AIR -filter:retweets"
date_since = "2020-11-10"
# Call the function scraptweets
scrape_tweets(search_words, date_since)

search_words = "#Apple chip -filter:retweets"
date_since = "2020-11-10"
# Call the function scraptweets
scrape_tweets(search_words, date_since)
